requirements.txt
streamlit>=1.28.0
langchain>=0.0.350
langchain-community>=0.0.10
langchain-google-genai>=0.0.6
pypdf2>=3.0.1
faiss-cpu>=1.8.0
sentence-transformers>=2.2.2
google-generativeai>=0.3.0
python-dotenv>=1.0.0
streamlit-chat>=0.1.1
streamlit-option-menu>=0.3.6
plotly>=5.17.0
pandas>=2.1.0
numpy>=1.24.0
